+++ packages/d3-state-visualizer/src/index.ts
@@ -1,2 +1,2 @@
 export { tree } from './charts';
-export type { InputOptions, NodeWithId, Primitive } from './charts';
+export type { HierarchyPointNode, Node, Options, StyleValue } from './charts';