+++ packages/redux-devtools-app/webpack.config.ts
@@ -39,9 +39,6 @@ export default (
   },
   resolve: {
     extensions: ['.js', '.jsx', '.ts', '.tsx'],
-    fallback: {
-      path: require.resolve('path-browserify'),
-    },
   },
   plugins: [
     new webpack.DefinePlugin({