+++ packages/redux-devtools-ui/src/Tabs/Tabs.tsx
@@ -14,7 +14,7 @@ export interface TabsProps<P> {
   position: Position;
 }
 
-export default class Tabs<P> extends Component<TabsProps<P>> {
+export default class Tabs<P extends object> extends Component<TabsProps<P>> {
   SelectedComponent?: React.ComponentType<P>;
   selector?: () => P;